+++ src/util/matrix4x4.cpp
@@ -8,9 +8,9 @@ using namespace std;
 
 namespace CGL {
 
-double &Matrix4x4::operator()(int i, int j) { return entries[j][i]; }
+float &Matrix4x4::operator()(int i, int j) { return entries[j][i]; }
 
-const double &Matrix4x4::operator()(int i, int j) const {
+const float &Matrix4x4::operator()(int i, int j) const {
   return entries[j][i];
 }
 
@@ -18,7 +18,7 @@ Vector4D &Matrix4x4::operator[](int j) { return entries[j]; }
 
 const Vector4D &Matrix4x4::operator[](int j) const { return entries[j]; }
 
-void Matrix4x4::zero(double val) {
+void Matrix4x4::zero(float val) {
   // sets all elements to val
   entries[0] = Vector4D(val);
   entries[1] = Vector4D(val);
@@ -26,7 +26,7 @@ void Matrix4x4::zero(double val) {
   entries[3] = Vector4D(val);
 }
 
-double Matrix4x4::det(void) const {
+float Matrix4x4::det(void) const {
   const Matrix4x4 &A(*this);
 
   return A(0, 3) * A(1, 2) * A(2, 1) * A(3, 0) -
@@ -55,8 +55,8 @@ double Matrix4x4::det(void) const {
          A(0, 0) * A(1, 1) * A(2, 2) * A(3, 3);
 }
 
-double Matrix4x4::norm(void) const {
-  return sqrt(entries[0].norm2() + entries[1].norm2() + entries[2].norm2() +
+float Matrix4x4::norm(void) const {
+  return sqrtf(entries[0].norm2() + entries[1].norm2() + entries[2].norm2() +
               entries[3].norm2());
 }
 
@@ -76,8 +76,8 @@ Matrix4x4 Matrix4x4::operator-(void) const {
 void Matrix4x4::operator+=(const Matrix4x4 &B) {
 
   Matrix4x4 &A(*this);
-  double *Aij = (double *)&A;
-  const double *Bij = (const double *)&B;
+  float *Aij = (float *)&A;
+  const float *Bij = (const float *)&B;
 
   A[0] += B[0];
   A[1] += B[1];
@@ -97,7 +97,7 @@ Matrix4x4 Matrix4x4::operator-(const Matrix4x4 &B) const {
   return C;
 }
 
-Matrix4x4 Matrix4x4::operator*(double c) const {
+Matrix4x4 Matrix4x4::operator*(float c) const {
   const Matrix4x4 &A(*this);
   Matrix4x4 B;
 
@@ -110,7 +110,7 @@ Matrix4x4 Matrix4x4::operator*(double c) const {
 }
 
 // Returns c*A.
-Matrix4x4 operator*(double c, const Matrix4x4 &A) {
+Matrix4x4 operator*(float c, const Matrix4x4 &A) {
   Matrix4x4 cA;
 
   cA[0] = c * A[0];
@@ -128,7 +128,7 @@ Matrix4x4 Matrix4x4::operator*(const Matrix4x4 &B) const {
 
   for (int i = 0; i < 4; i++) {
     for (int j = 0; j < 4; j++) {
-      C(i, j) = 0.;
+      C(i, j) = 0.f;
       for (int k = 0; k < 4; k++) {
         C(i, j) += A(i, k) * B(k, j);
       }
@@ -217,9 +217,9 @@ Matrix4x4 Matrix4x4::inv(void) const {
   return B;
 }
 
-void Matrix4x4::operator/=(double x) {
+void Matrix4x4::operator/=(float x) {
   Matrix4x4 &A(*this);
-  double rx = 1. / x;
+  float rx = 1.f / x;
 
   for (int i = 0; i < 4; i++)
     for (int j = 0; j < 4; j++) {
@@ -230,22 +230,22 @@ void Matrix4x4::operator/=(double x) {
 Matrix4x4 Matrix4x4::identity(void) {
   Matrix4x4 B;
 
-  B(0, 0) = 1.;
-  B(0, 1) = 0.;
-  B(0, 2) = 0.;
-  B(0, 3) = 0.;
-  B(1, 0) = 0.;
-  B(1, 1) = 1.;
-  B(1, 2) = 0.;
-  B(1, 3) = 0.;
-  B(2, 0) = 0.;
-  B(2, 1) = 0.;
-  B(2, 2) = 1.;
-  B(2, 3) = 0.;
-  B(3, 0) = 0.;
-  B(3, 1) = 0.;
-  B(3, 2) = 0.;
-  B(3, 3) = 1.;
+  B(0, 0) = 1.f;
+  B(0, 1) = 0.f;
+  B(0, 2) = 0.f;
+  B(0, 3) = 0.f;
+  B(1, 0) = 0.f;
+  B(1, 1) = 1.f;
+  B(1, 2) = 0.f;
+  B(1, 3) = 0.f;
+  B(2, 0) = 0.f;
+  B(2, 1) = 0.f;
+  B(2, 2) = 1.f;
+  B(2, 3) = 0.f;
+  B(3, 0) = 0.f;
+  B(3, 1) = 0.f;
+  B(3, 2) = 0.f;
+  B(3, 3) = 1.f;
 
   return B;
 }
